+++ pyo3-macros-backend/src/method.rs
@@ -471,7 +471,7 @@ impl<'a> FnSpec<'a> {
         let rust_call = |args: Vec<TokenStream>| {
             quote! {
                 let mut ret = #rust_name(#self_arg #(#args),*);
-                let owned = _pyo3::callback::OkWrap::wrap(ret, #py);
+                let owned = _pyo3::impl_::pymethods::OkWrap::wrap(ret, #py);
                 owned.map(|obj| _pyo3::conversion::IntoPyPointer::into_ptr(obj))
                     .map_err(::core::convert::Into::into)
             }
