+++ pyo3-macros-backend/src/pymethod.rs
@@ -389,7 +389,7 @@ fn impl_py_class_attribute(cls: &syn::Type, spec: &FnSpec<'_>) -> syn::Result<Me
         fn #wrapper_ident(py: _pyo3::Python<'_>) -> _pyo3::PyResult<_pyo3::PyObject> {
             #deprecations
             let mut ret = #fncall;
-            let owned = _pyo3::callback::OkWrap::wrap(ret, py);
+            let owned = _pyo3::impl_::pymethods::OkWrap::wrap(ret, py);
             owned.map_err(::core::convert::Into::into)
         }
     };
